requirements.txt
requests>=2.31.0
colorama>=0.4.6
customtkinter>=5.2.0
anthropic>=0.25.0
tenacity>=8.2.3
pydantic>=2.5.0
python-dotenv>=1.0.0
